spltis up Slot test files
