+++ test/slots/IntegerInSlot.test.js
@@ -1,0 +1,108 @@
+const IntegerInSlot = require('slots/IntegerInSlot')
+const { InSlotSeeder } = require('seeders/slots')
+
+describe('the IntegerInSlot class', () => {
+  describe('.constructor', () => {
+    const requiredKeys = [ 'defaultValue', 'min', 'max' ]
+    const forbiddenKeys = ['selectOptions']
+
+    it('creates a integer slot with the expected properties', () => {
+      const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
+      const slot = new IntegerInSlot(slotData)
+
+      requiredKeys.forEach(key => expect(key in slot).toBe(true))
+      forbiddenKeys.forEach(key => expect(key in slot).toBe(false))
+    })
+
+    it('sets defaultValue per default to null', () => {
+      const slotData = InSlotSeeder.generate({
+        dataType: IntegerInSlot.DATA_TYPE,
+      })
+      delete slotData.defaultValue
+
+      const slot = new IntegerInSlot(slotData)
+
+      expect(slot.defaultValue).toBeNull()
+    })
+
+    it('sets min per default to null', () => {
+      const slotData = InSlotSeeder.generate({
+        dataType: IntegerInSlot.DATA_TYPE,
+      })
+      delete slotData.min
+
+      const slot = new IntegerInSlot(slotData)
+
+      expect(slot.min).toBeNull()
+    })
+
+    it('sets max per default to null', () => {
+      const slotData = InSlotSeeder.generate({
+        dataType: IntegerInSlot.DATA_TYPE,
+      })
+      delete slotData.max
+
+      const slot = new IntegerInSlot(slotData)
+
+      expect(slot.max).toBeNull()
+    })
+
+    it('allows min and max to be null', () => {
+      const slotData = InSlotSeeder.generate({
+        dataType: IntegerInSlot.DATA_TYPE,
+      })
+      delete slotData.max
+      delete slotData.min
+
+      const slot = new IntegerInSlot(slotData)
+
+      expect(slot.max).toBeNull()
+      expect(slot.min).toBeNull()
+    })
+
+    it('throws when max is below min', () => {
+      const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
+      slotData.max = 100
+      slotData.min = 200
+
+      expect(() => new IntegerInSlot(slotData)).toThrow(/max/)
+    })
+
+    it('throws when defaultValue is above max', () => {
+      const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
+      slotData.max = 100
+      slotData.defaultValue = 200
+
+      expect(() => new IntegerInSlot(slotData)).toThrow(/max/)
+    })
+
+    it('throws when defaultValue is below min', () => {
+      const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
+      slotData.min = 300
+      slotData.defaultValue = 200
+
+      expect(() => new IntegerInSlot(slotData)).toThrow(/min/)
+    })
+
+    it('throws when defaultValue is a float number', () => {
+      const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
+      slotData.defaultValue = 200.1
+
+      expect(() => new IntegerInSlot(slotData)).toThrow(/defaultValue/)
+    })
+
+    it('throws when min is a float number', () => {
+      const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
+      slotData.min = 200.1
+
+      expect(() => new IntegerInSlot(slotData)).toThrow(/min/)
+    })
+
+    it('throws when max is a float number', () => {
+      const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
+      slotData.max = 200.1
+
+      expect(() => new IntegerInSlot(slotData)).toThrow(/max/)
+    })
+  })
+})
