+++ test/slots/Slot.test.js
@@ -3,10 +3,7 @@ const faker = require('faker')
 
 const Slot = require('slots/Slot')
 const InSlot = require('slots/InSlot')
-const BooleanInSlot = require('slots/BooleanInSlot')
-const OneOfInSlot = require('slots/OneOfInSlot')
 const IntegerInSlot = require('slots/IntegerInSlot')
-const FloatInSlot = require('slots/FloatInSlot')
 const OutSlot = require('slots/OutSlot')
 const DataStream = require('dataStreams/DataStream')
 const DataStreamSeeder = require('seeders/dataStreams/DataStreamSeeder')
@@ -41,211 +38,6 @@ describe('the Slot class', () => {
 
         expect(slot.tareable).toBe(false)
       })
-
-      describe('for a boolean dataType', () => {
-        const requiredKeys = [ 'defaultValue' ]
-        const forbiddenKeys = ['min', 'max', 'selectOptions']
-
-        it('creates a boolean slot with the expected properties', () => {
-          const slotData = InSlotSeeder.generate({ dataType: BooleanInSlot.DATA_TYPE })
-
-          const slot = new BooleanInSlot(slotData)
-
-          requiredKeys.forEach(key => expect(key in slot).toBe(true))
-          forbiddenKeys.forEach(key => expect(key in slot).toBe(false))
-        })
-      })
-
-      describe('for a oneOf dataType', () => {
-        const requiredKeys = [ 'defaultValue', 'selectOptions' ]
-        const forbiddenKeys = ['min', 'max']
-
-        it('creates a oneOf slot with the expected properties', () => {
-          const slotData = InSlotSeeder.generate({ dataType: OneOfInSlot.DATA_TYPE })
-          const slot = new OneOfInSlot(slotData)
-
-          requiredKeys.forEach(key => expect(key in slot).toBe(true))
-          forbiddenKeys.forEach(key => expect(key in slot).toBe(false))
-        })
-      })
-
-      describe('for an float dataType', () => {
-        const requiredKeys = [ 'defaultValue', 'min', 'max' ]
-        const forbiddenKeys = ['selectOptions']
-
-        it('creates a float slot with the expected properties', () => {
-          const slotData = InSlotSeeder.generate({ dataType: FloatInSlot.DATA_TYPE })
-          const slot = new FloatInSlot(slotData)
-
-          requiredKeys.forEach(key => expect(key in slot).toBe(true))
-          forbiddenKeys.forEach(key => expect(key in slot).toBe(false))
-        })
-
-        it('sets min per default to null', () => {
-          const slotData = InSlotSeeder.generate({
-            dataType: FloatInSlot.DATA_TYPE,
-          })
-
-          delete slotData.min
-
-          const slot = new FloatInSlot(slotData)
-
-          expect(slot.min).toBeNull()
-        })
-
-        it('sets max per default to null', () => {
-          const slotData = InSlotSeeder.generate({
-            dataType: FloatInSlot.DATA_TYPE,
-          })
-
-          delete slotData.max
-
-          const slot = new FloatInSlot(slotData)
-
-          expect(slot.max).toBeNull()
-        })
-
-        it('allows min and max to be null', () => {
-          const slotData = InSlotSeeder.generate({
-            dataType: FloatInSlot.DATA_TYPE,
-          })
-
-          delete slotData.max
-          delete slotData.min
-
-          const slot = new FloatInSlot(slotData)
-
-          expect(slot.max).toBeNull()
-          expect(slot.min).toBeNull()
-        })
-
-        it('throws when max is below min', () => {
-          const slotData = InSlotSeeder.generate({ dataType: FloatInSlot.DATA_TYPE })
-          slotData.max = 100.1
-          slotData.min = 200.2
-
-          expect(() => new FloatInSlot(slotData)).toThrow(/max/)
-        })
-
-        it('throws when defaultValue is above max', () => {
-          const slotData = InSlotSeeder.generate({ dataType: FloatInSlot.DATA_TYPE })
-          slotData.max = 100.1
-          slotData.defaultValue = 200.1
-
-          expect(() => new FloatInSlot(slotData)).toThrow(/max/)
-        })
-
-        it('throws when defaultValue is below min', () => {
-          const slotData = InSlotSeeder.generate({ dataType: FloatInSlot.DATA_TYPE })
-          slotData.min = 300.1
-          slotData.defaultValue = 200.1
-
-          expect(() => new FloatInSlot(slotData)).toThrow(/min/)
-        })
-      })
-
-      describe('for an integer dataType', () => {
-        const requiredKeys = [ 'defaultValue', 'min', 'max' ]
-        const forbiddenKeys = ['selectOptions']
-
-        it('creates a integer slot with the expected properties', () => {
-          const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
-          const slot = new IntegerInSlot(slotData)
-
-          requiredKeys.forEach(key => expect(key in slot).toBe(true))
-          forbiddenKeys.forEach(key => expect(key in slot).toBe(false))
-        })
-
-        it('sets min per default to null', () => {
-          const slotData = InSlotSeeder.generate({
-            dataType: IntegerInSlot.DATA_TYPE,
-          })
-
-          delete slotData.min
-
-          const slot = new IntegerInSlot(slotData)
-
-          expect(slot.min).toBeNull()
-        })
-
-        it('sets max per default to null', () => {
-          const slotData = InSlotSeeder.generate({
-            dataType: IntegerInSlot.DATA_TYPE,
-          })
-
-          delete slotData.max
-
-          const slot = new IntegerInSlot(slotData)
-
-          expect(slot.max).toBeNull()
-        })
-
-        it('allows min and max to be null', () => {
-          const slotData = InSlotSeeder.generate({
-            dataType: IntegerInSlot.DATA_TYPE,
-          })
-
-          delete slotData.max
-          delete slotData.min
-
-          const slot = new IntegerInSlot(slotData)
-
-          expect(slot.max).toBeNull()
-          expect(slot.min).toBeNull()
-        })
-
-        it('throws when max is below min', () => {
-          const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
-          slotData.max = 100
-          slotData.min = 200
-
-          expect(() => new IntegerInSlot(slotData)).toThrow(/max/)
-        })
-
-        it('throws when defaultValue is above max', () => {
-          const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
-          slotData.max = 100
-          slotData.defaultValue = 200
-
-          expect(() => new FloatInSlot(slotData)).toThrow(/max/)
-        })
-
-        it('throws when defaultValue is below min', () => {
-          const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
-          slotData.min = 300
-          slotData.defaultValue = 200
-
-          expect(() => new IntegerInSlot(slotData)).toThrow(/min/)
-        })
-
-        it('throws when defaultValue is a float number', () => {
-          const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
-          slotData.defaultValue = 200.1
-
-          expect(() => new IntegerInSlot(slotData)).toThrow(/defaultValue/)
-        })
-
-        it('throws when min is a float number', () => {
-          const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
-          slotData.min = 200.1
-
-          expect(() => new IntegerInSlot(slotData)).toThrow(/min/)
-        })
-
-        it('throws when max is a float number', () => {
-          const slotData = InSlotSeeder.generate({ dataType: IntegerInSlot.DATA_TYPE })
-          slotData.max = 200.1
-
-          expect(() => new IntegerInSlot(slotData)).toThrow(/max/)
-        })
-      })
-    })
-
-    it('allows null for displayType', () => {
-      const slotData = SlotSeeder.generate({ displayType: null })
-      const slot = new Slot(slotData)
-
-      expect(slot.displayType).toBeNull()
     })
 
     it('converts dataStream data in to DataStream instances', () => {
